VetDataHub/requirements.txt
dj-database-url==0.4.1
Django==1.10.4
gunicorn==19.6.0
psycopg2
whitenoise==3.2
Pillow
